consume libhoney responses channel in a thread, preventing a deadlock
We found that libhoney provides a responses() channel which MUST be read to
prevent eventual deadlock of the process. libhoney-rust writes to the responses
channel for every event written to honeycomb. This eventually causes the
work_sender thread to block, which causes the trace-data sending thread to
block.

We now consume the honeycomb responses in a tight loop, eventually exiting
if we detect that the channel has disconnected and all messages have been
consumed (the `responses.recv.is_err()` case.)